services/voice-processing/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
boto3==1.34.25
aioboto3==12.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
numpy==1.26.3
